+++ src/SolutionInfo.cs
@@ -6,8 +6,8 @@
 using System.Reflection;
 
 [assembly: AssemblyProduct("Cake IIS")]
-[assembly: AssemblyVersion("0.0.6")]
-[assembly: AssemblyFileVersion("0.0.6")]
-[assembly: AssemblyInformationalVersion("0.0.6")]
+[assembly: AssemblyVersion("0.0.7")]
+[assembly: AssemblyFileVersion("0.0.7")]
+[assembly: AssemblyInformationalVersion("0.0.7")]
 [assembly: AssemblyCopyright("Copyright (c) Sergio Silveira, Phillip Sharpe 2015")]